+++ apps/vs-code-designer/src/app/utils/dotnet/executeDotnetTemplateCommand.ts
@@ -3,12 +3,44 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 import { localize } from '../../../localize';
-import { executeCommand } from '../funcCoreTools/cpUtils';
+import { executeCommand, wrapArgInQuotes } from '../funcCoreTools/cpUtils';
+import { ext } from '@microsoft/vscode-azext-azureappservice/out/src/extensionVariables';
 import type { IActionContext } from '@microsoft/vscode-azext-utils';
+import type { FuncVersion } from '@microsoft/vscode-extension';
+import path from 'path';
 import type { SemVer } from 'semver';
 import { coerce as semVerCoerce } from 'semver';
 
 let cachedFramework: string | undefined;
+
+export async function executeDotnetTemplateCommand(
+  context: IActionContext,
+  version: FuncVersion,
+  projTemplateKey: string,
+  workingDirectory: string | undefined,
+  operation: 'list' | 'create',
+  ...args: string[]
+): Promise<string> {
+  const framework: string = await getFramework(context, workingDirectory);
+  const jsonDllPath: string = ext.context.asAbsolutePath(
+    path.join('resources', 'dotnetJsonCli', framework, 'Microsoft.TemplateEngine.JsonCli.dll')
+  );
+  return await executeCommand(
+    undefined,
+    workingDirectory,
+    'dotnet',
+    wrapArgInQuotes(jsonDllPath),
+    '--templateDir',
+    wrapArgInQuotes(getDotnetTemplateDir(version, projTemplateKey)),
+    '--operation',
+    operation,
+    ...args
+  );
+}
+
+export function getDotnetTemplateDir(version: FuncVersion, projTemplateKey: string): string {
+  return path.join(ext.context.globalStoragePath, '', version, projTemplateKey);
+}
 
 /**
  * Validates .NET is installed.
