+++ apps/vs-code-designer/src/constants.ts
@@ -27,6 +27,7 @@ export const hostStartCommand = 'host start';
 export const hostStartTaskName = `${func}: ${hostStartCommand}`;
 export const funcPackageName = 'azure-functions-core-tools';
 export const defaultFuncPort = '7071';
+export const isolatedSdkName = 'Microsoft.Azure.Functions.Worker.Sdk';
 
 // Workflow
 export const workflowLocationKey = 'WORKFLOWS_LOCATION_NAME';
