--- libs/vscode-extension/src/lib/models/templates/IFunctionTemplate.ts
@@ -1,20 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-import type { TemplateCategory } from '..';
-import type { IBindingSetting } from './IBindingTemplate';
-
-/**
- * Describes a template used for creating a function trigger (i.e. an HttpTrigger or TimerTrigger)
- */
-export interface IFunctionTemplate {
-  id: string;
-  name: string;
-  defaultFunctionName: string;
-  language: string;
-  isHttpTrigger: boolean;
-  isTimerTrigger: boolean;
-  userPromptedSettings: IBindingSetting[];
-  categories: TemplateCategory[];
-}
